+++ CHANGELOG.md
@@ -6,17 +6,17 @@ v0.0.4 (????-??-??)
 -------------------
 - Clean up code by extracting out complex_type_start
 - Extract out valid extensions to an array
+- Allow disabling of the file extension check
 
 v0.0.3 (2015-01-15)
 -------------------
 - **[ISSUE #1](https://github.com/shortdudey123/yamllint/issues/1)** - Add more verbose output
-- **[ISSUE #2](https://github.com/shortdudey123/yamllint/issues/2)** - Add file extention verification
+- **[ISSUE #2](https://github.com/shortdudey123/yamllint/issues/2)** - Add file extension verification
 - add fail_on_error option to rake task
 
 v0.0.2 (2015-01-15)
 -------------------
 - Fix rake_task filename
-
 
 v0.0.1 (2015-01-15)
 -------------------